requirements.txt
discord.py>=2.3.2
openai>=1.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
flask>=2.3.0
flask-cors>=4.0.0
asyncio
google-generativeai>=0.7.0
